requirements.txt
fastapi==0.111.1
pydantic==2.8.2
pydantic[email]
psycopg2-binary==2.9.9
databases==0.9.0
sqlalchemy==2.0.31
pyodbc==5.1.0
uvicorn==0.30.3
python-dotenv==1.0.1
pytz==2024.1
pytest==8.3.3
pytest-mock==3.14.0
httpx